Add @kansa/common/split-name & use it for badge names
